requirements.txt
streamlit
joblib
pandas
scikit-learn
mlflow
kafka-python
requests
fpdf2
python-dateutil
numpy
sqlite3
matplotlib
seaborn
